% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixAll.R
\docType{data}
\name{birds}
\alias{birds}
\title{Qualitative data : morphological description of birds}
\format{A data frame with 69 observations on the following 5 variables.

\describe{
  \item{\code{gender}}{a character vector defining the gender (2 modalities, male or female).}
  \item{\code{eyebrow}}{a character vector describing the eyebrow stripe (4 modalities).}
  \item{\code{collar}}{a character vector describing the collar (5 modalities).}
  \item{\code{sub-caudal}}{a character vector describing the sub-caudal (5 modalities).}
  \item{\code{border}}{a character vector describing the border (3 modalities).}
}}
\source{
Bretagnolle, V., 2007. Personal communication, source: Museum.
}
\description{
The data set contains details on the morphology of birds (puffins). Each
individual (bird) is described by 6 qualitative variables. One variable for
the gender and 5 variables giving a morphological description of the birds.
There is 69 puffins divided in 2 sub-classes: lherminieri (34) and subalaris (35).
}
\details{
This data set is also part of the Rmixmod package.
}
\examples{
  data(birds)
}
\keyword{datasets}
