% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixAll.R
\docType{data}
\name{bullsEye.cat}
\alias{bullsEye.cat}
\title{label Data: bullsEye.cat}
\description{
Generated data set containing two categorical variables for the two clusters
with untypical ring shapes (circles)
}
\examples{
  data(bullsEye.cat)
}
\keyword{datasets}
