% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixAll.R
\docType{data}
\name{bullsEye}
\alias{bullsEye}
\title{Quantitative Data: bullsEye}
\description{
Generated data set containing two clusters with untypical ring shapes (circles)
}
\examples{
  data(bullsEye)
}
\keyword{datasets}
