% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixAll.R
\docType{data}
\name{bullsEye.target}
\alias{bullsEye.target}
\title{label Data: bullsEye.target}
\description{
Generated data set containing labels for the two clusters with untypical
ring shapes (circles)
}
\examples{
  data(bullsEye.target)
}
\keyword{datasets}
