% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAlgoPredict.R
\docType{class}
\name{ClusterAlgoPredict}
\alias{ClusterAlgoPredict}
\alias{ClusterAlgoPredict-class}
\title{[\code{\linkS4class{ClusterAlgoPredict}}] class for predict algorithm.}
\description{
This class encapsulates the parameters of prediction methods.
}
\section{Slots}{

\describe{
\item{\code{algo}}{A character string with the algorithm.
Possible values: "EM", "SemiSEM". Default value: "SemiSEM".}

\item{\code{nbIterBurn}}{Integer defining the number of burning iterations. Default value is 50.}

\item{\code{nbIterLong}}{Integer defining the number of iterations. Default value is 100.}

\item{\code{epsilon}}{real defining the epsilon value for the long algorithm. epsilon is
note used if \code{algo} is "SemiSEM". Default value: 1e-07.}
}}

\examples{
getSlots("ClusterAlgoPredict")
new("ClusterAlgoPredict")
new("ClusterAlgoPredict", algo="SemiSEM", nbIterBurn=10)

}
