% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterCategorical.R
\name{clusterCategorical}
\alias{clusterCategorical}
\title{Create an instance of the [\code{\linkS4class{ClusterCategorical}}] class}
\usage{
clusterCategorical(data, nbCluster = 2,
  models = clusterCategoricalNames(probabilities = "free"),
  strategy = clusterStrategy(), criterion = "ICL", nbCore = 1)
}
\arguments{
\item{data}{a data.frame or a matrix containing the data. Rows correspond to
observations and columns correspond to variables. data will be coerced as
an integer matrix. If data set contains NA values, they will be estimated
during the estimation process.}

\item{nbCluster}{[\code{\link{vector}}] listing the number of clusters to test.}

\item{models}{[\code{\link{vector}}] of model names to run. By default
the categorical models "categorical_pk_pjk" and "categorical_p_pjk" are estimated.}

\item{strategy}{a [\code{\linkS4class{ClusterStrategy}}] object containing
the strategy to run. [\code{\link{clusterStrategy}}]() method by default.}

\item{criterion}{character defining the criterion to select the best model.
The best model is the one with the lowest criterion value.
Possible values: "BIC", "AIC", "ICL", "ML". Default is "ICL".}

\item{nbCore}{integer defining the number of processors to use (default is 1, 0 for all).}
}
\value{
An instance of the [\code{\linkS4class{ClusterCategorical}}] class.
}
\description{
This function computes the optimal Categorical mixture model according
to the \code{criterion} among the list of model given in \code{models}
and the number of clusters given in \code{nbCluster}, using the strategy
specified in \code{strategy}.
}
\examples{
## A quantitative example with the birds data set
data(birds)
## add 10 missing values
x = as.matrix(birds); n <- nrow(x); p <- ncol(x)
indexes <- matrix(c(round(runif(5,1,n)), round(runif(5,1,p))), ncol=2)
x[indexes] <- NA
## estimate model (using fast strategy, results may be misleading)
model <- clusterCategorical( data=x, nbCluster=2:3
                           , models=c( "categorical_pk_pjk", "categorical_p_pjk")
                           , strategy = clusterFastStrategy()
                           )

## use graphics functions
\dontrun{
plot(model)
}

## get summary
summary(model)
## print model
\dontrun{
print(model)
}
## get estimated missing values
missingValues(model)

}
\author{
Serge Iovleff
}
