% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterCategorical.R
\docType{class}
\name{ClusterCategoricalComponent}
\alias{ClusterCategoricalComponent}
\alias{ClusterCategoricalComponent-class}
\title{Definition of the [\code{\linkS4class{ClusterCategoricalComponent}}] class}
\description{
This class defines a categorical component of a mixture model. It inherits
from [\code{\linkS4class{IClusterComponent}}].
}
\section{Slots}{

\describe{
\item{\code{plkj}}{Array with the probability for the jth variable in the kth cluster to be l.}

\item{\code{nbModalities}}{Integer with the (maximal) number of modalities of the categorical data.}

\item{\code{levels}}{list with the original levels of the variables}
}}

\examples{
getSlots("ClusterCategoricalComponent")

}
\seealso{
[\code{\linkS4class{IClusterComponent}}] class
}
\author{
Serge Iovleff
}
