% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDiagGaussian.R
\docType{class}
\name{ClusterDiagGaussianComponent}
\alias{ClusterDiagGaussianComponent}
\alias{ClusterDiagGaussianComponent-class}
\title{Definition of the [\code{\linkS4class{ClusterDiagGaussianComponent}}] class}
\description{
This class defines a diagonal Gaussian component of a mixture model. It inherits
from [\code{\linkS4class{IClusterComponent}}].
}
\section{Slots}{

\describe{
\item{\code{mean}}{Matrix with the mean of the jth variable in the kth cluster.}

\item{\code{sigma}}{Matrix with the standard deviation of the jth variable in the kth cluster.}
}}

\examples{
getSlots("ClusterDiagGaussianComponent")

}
\seealso{
[\code{\linkS4class{IClusterComponent}}] class
}
\author{
Serge Iovleff
}
