% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClusterGamma.R
\docType{class}
\name{ClusterGammaComponent}
\alias{ClusterGammaComponent}
\alias{ClusterGammaComponent-class}
\title{Definition of the [\code{\linkS4class{ClusterGammaComponent}}] class}
\description{
This class defines a gamma component of a mixture Model. It inherits
from [\code{\linkS4class{IClusterComponent}}].
}
\section{Slots}{

\describe{
\item{\code{shape}}{Matrix with the shape of the jth variable in the kth cluster.}

\item{\code{scale}}{Matrix with the scale of the jth variable in the kth cluster.}
}}
\examples{
getSlots("ClusterGammaComponent")
}
\author{
Serge Iovleff
}
\seealso{
[\code{\linkS4class{IClusterComponent}}] class
}

