% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterKernel.R
\docType{class}
\name{ClusterKernelComponent}
\alias{ClusterKernelComponent}
\alias{ClusterKernelComponent-class}
\title{Definition of the [\code{\linkS4class{ClusterKernelComponent}}] class}
\description{
This class defines a kernel component of a mixture Model. It inherits
from [\code{\linkS4class{IClusterComponent}}].
}
\section{Slots}{

\describe{
\item{\code{dim}}{Vector with the dimension of the kth cluster}

\item{\code{sigma}}{Vector with the standard deviation in the kth cluster.}

\item{\code{rawData}}{can be used to save the original data set as the slot data
from [\code{\linkS4class{IClusterComponent}}] will be overwrite by the gram
matrix.}
}}
\examples{
getSlots("ClusterKernelComponent")

}
\author{
Serge Iovleff
}
\seealso{
[\code{\linkS4class{IClusterComponent}}] class
}

