% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMixedData.R
\docType{class}
\name{ClusterMixedData}
\alias{ClusterMixedData}
\alias{ClusterMixedData-class}
\title{Definition of the [\code{\linkS4class{ClusterMixedData}}] class}
\description{
This class defines a mixed data mixture Model.
}
\details{
This class inherits from the [\code{\linkS4class{IClusterModel}}] class.
A model for mixed data is a mixture model of the form:
\deqn{
f({{x}}_i=({{x}}_{1i}, {{x}}_{2i},\ldots {{x}}_{Li})|\theta)
= \sum_{k=1}^K p_k \prod_{l=1}^L h({{x}}_{li}| \lambda_{lk},\alpha_l).
}
The density functions (or probability distribution functions)
\deqn{h(.|\lambda_{lk},\alpha_l)}
can be any implemented model (Gaussian, Poisson,...).
}
\section{Slots}{

\describe{
\item{\code{ldata}}{a list of IClusterComponent.}
}}
\examples{
getSlots("ClusterMixedData")

}
\author{
Serge Iovleff
}
\seealso{
[\code{\linkS4class{IClusterModel}}] class
}

