% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IClusterModel.R
\docType{class}
\name{IClusterModel}
\alias{IClusterModel}
\alias{IClusterModel-class}
\title{Interface base Class [\code{\linkS4class{IClusterModel}}] for Cluster models.}
\description{
This class encapsulate the common parameters of all the Cluster models.
}
\details{
A Cluster model is a model of the form
\deqn{
  f({x}|\boldsymbol{\theta})
    \sum_{k=1}^K p_k h({x};\boldsymbol{\lambda}_k,\boldsymbol{\alpha})
   \quad {x} \in J.
}
where h can be either a pdf, a discrete probability, (homogeneous case)
or a product of arbitrary pdf and discrete probabilities (mixed data case).
}
\section{Slots}{

\describe{
\item{\code{nbSample}}{Integer with the number of samples of the model.}

\item{\code{nbCluster}}{Integer with the number of cluster of the model.}

\item{\code{pk}}{Vector of size K with the proportions of each mixture.}

\item{\code{tik}}{Matrix of size \eqn{n \times K} with the posterior probability of
the ith individual to belong to kth cluster.}

\item{\code{lnFi}}{Vector of size n with the log-likelihood of the ith individuals.}

\item{\code{zi}}{Vector of integer of size n with the attributed class label of the individuals.}

\item{\code{ziFit}}{Vector of integer of size n with the fitted class label of the individuals (only used in supervised learning).}

\item{\code{lnLikelihood}}{Real given the ln-liklihood of the Cluster model.}

\item{\code{criterion}}{Real given the value of the AIC, BIC, ICL or ML criterion.}

\item{\code{criterionName}}{string with the name of the criterion. Possible values
are "BIC", "AIC", "ICL" or "ML". Default is "ICL".}

\item{\code{nbFreeParameter}}{Integer given the number of free parameters of the model.}

\item{\code{strategy}}{the instance of the [\code{\linkS4class{ClusterStrategy}}] used in the
estimation process of the mixture. Default is clusterStrategy().}
}}

\examples{
  getSlots("IClusterModel")

}
\author{
Serge Iovleff
}
