% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterPoisson.R
\docType{class}
\name{ClusterPoisson}
\alias{ClusterPoisson}
\alias{ClusterPoisson-class}
\title{Definition of the [\code{\linkS4class{ClusterPoisson}}] class}
\description{
This class inherits from the [\code{\linkS4class{IClusterModel}}] class.
A poisson mixture model is a mixture model of the form:
\deqn{
  f({x}|\boldsymbol{\theta}) \\
  =\sum_{k=1}^K p_k \prod_{j=1}^d Pois(x_j;\lambda_{jk}) \\
   \quad {x} \in {N}^d.
}
}
\section{Slots}{

\describe{
\item{\code{component}}{A [\code{\linkS4class{ClusterPoissonComponent}}] with the
lambda of the component mixture model.}
}}

\examples{
  getSlots("ClusterPoisson")
  data(DebTrivedi)
  dt <- DebTrivedi[, c(1, 6,8, 15)]
  new("ClusterPoisson", data=dt)

}
\seealso{
[\code{\linkS4class{IClusterModel}}] class
}
\author{
Serge Iovleff
}
