% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterPredict.R
\docType{class}
\name{ClusterPredict}
\alias{ClusterPredict}
\alias{ClusterPredict-class}
\title{Class [\code{\linkS4class{ClusterPredict}}] for predicting}
\description{
This class encapsulate the parameters for predicted data.
}
\section{Slots}{

\describe{
\item{\code{data}}{Matrix with the data set}

\item{\code{missing}}{Matrix with the indexes of the missing values}

\item{\code{nbSample}}{Integer with the number of samples}

\item{\code{nbCluster}}{Integer with the number of cluster}

\item{\code{pk}}{Vector of size K with the proportions of each mixture.}

\item{\code{tik}}{Matrix of size \eqn{n \times K} with the posterior probability of
the ith individual to belong to kth cluster.}

\item{\code{lnFi}}{Vector of size n with the log-likelihood of the ith individuals.}

\item{\code{zi}}{Vector of integer of size n  with the attributed class label of the individuals}

\item{\code{algo}}{an instance of [\code{\linkS4class{ClusterAlgoPredict}}]}
}}

\examples{
  getSlots("ClusterPredict")

}
\author{
Serge Iovleff
}
