% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterPredict.R
\docType{class}
\name{ClusterPredictMixedData}
\alias{ClusterPredictMixedData}
\alias{ClusterPredictMixedData-class}
\title{Class [\code{\linkS4class{ClusterPredictMixedData}}] for predicting}
\description{
This class encapsulate the parameters for predicted data.
}
\section{Slots}{

\describe{
\item{\code{ldata}}{list of matrix with the data sets}

\item{\code{lmissing}}{list of matrix with the indexes of the missing values}
}}

\examples{
  getSlots("ClusterPredictMixedData")

}
\seealso{
[\code{\linkS4class{IClusterPredict}}] class
}
\author{
Serge Iovleff
}
