% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixAll.R
\docType{data}
\name{DebTrivedi}
\alias{DebTrivedi}
\title{Counting Data: DebTrivedi}
\source{
\url{http://www.jstatsoft.org/v27/i08/paper}
}
\description{
Deb and Trivedi (1997) analyze data on 4406 individuals, aged 66 and over, who are covered
by Medicare, a public insurance program. Originally obtained from the US National Medical
Expenditure Survey (NMES) for 1987/88, the data are available from the data archive of
the \emph{Journal of Applied Econometrics}.
It was prepared for an R package accompanying Kleiber and Zeileis (2008)
and is also available asDebTrivedi.rda in the Journal of Statistical Software
together with Zeileis (2006). The objective is to model the demand for medical
care -as captured by the number of physician/non-physician office and hospital
outpatient visits- by the covariates available for the patients.
}
\examples{
  data(DebTrivedi)
}
\references{
Zeileis, A. and Kleiber, C. and Jackma, S. (2008). "Regression Models for Count Data in R". JSS 27, 8, 1--25.
}
\keyword{datasets}
