% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAlgo.R, R/ClusterAlgoPredict.R,
%   R/ClusterInit.R, R/ClusterStrategy.R, R/global.R, R/ClusterCategorical.R,
%   R/ClusterDiagGaussian.R, R/ClusterGamma.R, R/ClusterPoisson.R,
%   R/LearnAlgo.R, R/kmm.R
\docType{methods}
\name{[,ClusterAlgo-method}
\alias{[,ClusterAlgo-method}
\alias{[}
\alias{[<-,ClusterAlgo-method}
\alias{[,ClusterAlgoPredict-method}
\alias{[<-,ClusterAlgoPredict-method}
\alias{[,ClusterInit-method}
\alias{[<-,ClusterInit-method}
\alias{[,ClusterStrategy-method}
\alias{[<-,ClusterStrategy-method}
\alias{Extract}
\alias{[,ClusterCategoricalComponent-method}
\alias{[,ClusterDiagGaussianComponent-method}
\alias{[,ClusterGammaComponent-method}
\alias{[,ClusterPoissonComponent-method}
\alias{[,LearnAlgo-method}
\alias{[<-,LearnAlgo-method}
\alias{[,KmmComponent-method}
\title{Extract parts of a MixAll S4 class}
\usage{
\S4method{[}{ClusterAlgo}(x, i, j, drop)

\S4method{[}{ClusterAlgo}(x, i, j) <- value

\S4method{[}{ClusterAlgoPredict}(x, i, j, drop)

\S4method{[}{ClusterAlgoPredict}(x, i, j) <- value

\S4method{[}{ClusterInit}(x, i, j, drop)

\S4method{[}{ClusterInit}(x, i, j) <- value

\S4method{[}{ClusterStrategy}(x, i, j, drop)

\S4method{[}{ClusterStrategy}(x, i, j) <- value

\S4method{[}{ClusterCategoricalComponent}(x, i, j, drop)

\S4method{[}{ClusterDiagGaussianComponent}(x, i, j, drop)

\S4method{[}{ClusterGammaComponent}(x, i, j, drop)

\S4method{[}{ClusterPoissonComponent}(x, i, j, drop)

\S4method{[}{LearnAlgo}(x, i, j, drop)

\S4method{[}{LearnAlgo}(x, i, j) <- value

\S4method{[}{KmmComponent}(x, i, j, drop)
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s).}

\item{i}{the name of the element we want to extract or replace.}

\item{j}{if the element designing by i is complex, j specifying elements to extract or replace.}

\item{drop}{For matrices and arrays.  If TRUE the result is coerced to the lowest
possible dimension (see the examples).  This only works for extracting elements,
not for the replacement.  See drop for further details.}

\item{value}{typically an array-like R object of a similar class as the element
of x we want to replace.}
}
\description{
Extract parts of a MixAll S4 class
}
