% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClusterPoisson.R, R/IClusterModel.R
\docType{class}
\name{ClusterPoissonComponent}
\alias{ClusterPoissonComponent}
\alias{ClusterPoissonComponent-class}
\alias{IClusterComponent}
\alias{IClusterComponent-class}
\title{Definition of the [\code{\linkS4class{ClusterPoissonComponent}}] class}
\description{
This class defines a poisson component of a mixture Model. It inherits
from [\code{\linkS4class{IClusterComponent}}].

This class defines a component of a mixture Model
}
\section{Slots}{

\describe{
\item{\code{lambda}}{Matrix with the mean of the jth variable in the kth cluster.}

\item{\code{data}}{Matrix with the data set}

\item{\code{missing}}{Matrix with the indexes of the missing values}

\item{\code{modelName}}{model name associated with the data set}
}}
\examples{
getSlots("ClusterPoissonComponent")
getSlots("IClusterComponent")
}
\author{
Serge Iovleff

Serge Iovleff
}
\seealso{
[\code{\linkS4class{IClusterComponent}}] class
}

