% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IClusterPredict.R
\docType{class}
\name{IClusterPredict}
\alias{IClusterPredict}
\alias{IClusterPredict-class}
\title{Interface class [\code{\linkS4class{IClusterPredict}}] for predicting}
\description{
Interface base class for predicting clusters
}
\section{Slots}{

\describe{
\item{\code{nbSample}}{Integer with the number of samples}

\item{\code{nbCluster}}{Integer with the number of cluster}

\item{\code{pk}}{Vector of size K with the proportions of each mixture.}

\item{\code{tik}}{Matrix of size \eqn{n \times K} with the posterior probability of
the ith individual to belong to kth cluster.}

\item{\code{lnFi}}{Vector of size n with the log-likelihood of the ith individuals.}

\item{\code{zi}}{Vector of integer of size n  with the attributed class label of the individuals}

\item{\code{algo}}{an instance of [\code{\linkS4class{ClusterAlgoPredict}}]}

\item{\code{model}}{an instance of a (derived) [\code{\linkS4class{IClusterModel}}]}
}}

\examples{
  getSlots("IClusterPredict")

}
\author{
Serge Iovleff
}
