% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmmNames.R
\name{kmmNames}
\alias{kmmNames}
\alias{kmmValidModelNames}
\alias{kmmValidKernelNames}
\title{Create a vector of Kernel mixture model (KMM) names.}
\usage{
kmmNames(sdBetweenCluster = "all")

kmmValidModelNames(names)

kmmValidKernelNames(names)
}
\arguments{
\item{sdBetweenCluster}{A character string equal to "equal", "free" or "all". Default is "all".}

\item{names}{a vector of character with the names to check}
}
\value{
A vector of character with the model names.

TRUE if the names in the vector names are valid, FALSE otherwise.
}
\description{
In a Kernel mixture model, we can assume that
\enumerate{
 \item {The standard deviations are equal or free for all the clusters.}
}
This give rise to two models.

check if a vector of kernel mixture model (KMM) name is correct.

check if a vector of kernel name is correct.
}
\details{
The model names are summarized in the following array:
\tabular{ll}{
 Model Name  \tab standard deviation between clusters \cr
 kmm_sk      \tab Free                                \cr
 kmm_s       \tab Equal                               \cr
}
}
\examples{
kmmNames()
## same as c("kmm_sk")
kmmNames( sdBetweenCluster= "free")

}
