% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnAlgo.R
\docType{class}
\name{LearnAlgo}
\alias{LearnAlgo}
\alias{LearnAlgo-class}
\title{[\code{\linkS4class{LearnAlgo}}] class for Cluster algorithms.}
\description{
This class encapsulates the parameters of clustering estimation algorithms
methods.
}
\section{Slots}{

\describe{
\item{\code{algo}}{A character string with the algorithm.
Possible values: "Simul", "Impute. Default value: "Simul".}

\item{\code{nbIteration}}{Integer defining the maximal number of iterations. Default value: 200.}

\item{\code{epsilon}}{real defining the epsilon value for the algorithm. epsilon is
note used if \code{algo} is "Simul". Default value: 1e-07.}
}}

\examples{
getSlots("LearnAlgo")
new("LearnAlgo")
new("LearnAlgo", algo="Impute", nbIteration=100)

}
