% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterCategorical.R
\docType{methods}
\name{plot,ClusterCategorical-method}
\alias{plot,ClusterCategorical-method}
\alias{plot-ClusterCategorical}
\title{Plotting of a class [\code{\linkS4class{ClusterCategorical}}]}
\usage{
\S4method{plot}{ClusterCategorical}(x, y, ...)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{ClusterCategorical}}]}

\item{y}{a number between 1 and K-1.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting data from a [\code{\linkS4class{ClusterCategorical}}] object
using the estimated parameters and partition.
}
\examples{
  ## the car data set
\dontrun{
  data(car)
  model <- clusterCategorical(car, 3, strategy = clusterFastStrategy())
  plot(model)
  }

}
\seealso{
\code{\link{plot}}
}
