% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClusterKernel.R
\docType{methods}
\name{plot,ClusterKernel-method}
\alias{plot,ClusterKernel-method}
\alias{plot-ClusterKernel}
\title{Plotting of a class [\code{\linkS4class{ClusterKernel}}]}
\usage{
\S4method{plot}{ClusterKernel}(x, y, ...)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{ClusterKernel}}]}

\item{y}{not used.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting data from a [\code{\linkS4class{ClusterKernel}}] object
using the estimated parameters and partition.
}
\examples{
## the famous iris data set
\dontrun{
  data(iris)
  model <- clusterKernel(iris[1:4], 3, strategy = clusterFastStrategy())
  plot(model)
  }
}
\seealso{
\code{\link{plot}}
}

