% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMixedData.R
\docType{methods}
\name{plot,ClusterMixedDataModel-method}
\alias{plot,ClusterMixedDataModel-method}
\alias{plot-ClusterMixedDataModel}
\title{Plotting of a class [\code{\linkS4class{ClusterMixedDataModel}}]}
\usage{
\S4method{plot}{ClusterMixedDataModel}(x, y, ...)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{ClusterMixedDataModel}}]}

\item{y}{a number between 1 and K-1.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting data from a [\code{\linkS4class{ClusterMixedDataModel}}] object
using the estimated parameters and partition.
}
\examples{
\dontrun{
  ## the car data set
  data(car)
  model <- clusterMixedData(car, 3, strategy = clusterFastStrategy())
  plot(model)
  }

}
\seealso{
\code{\link{plot}}
}
