% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmm.R
\docType{methods}
\name{plot,KernelMixtureModel-method}
\alias{plot,KernelMixtureModel-method}
\alias{plot-KernelMixtureModel}
\title{Plotting of a class [\code{\linkS4class{KernelMixtureModel}}]}
\usage{
\S4method{plot}{KernelMixtureModel}(x, y, ...)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{KernelMixtureModel}}]}

\item{y}{a list of variables to plot (subset). Variables names or indices.
If missing all the variables are represented.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting data from a [\code{\linkS4class{KernelMixtureModel}}] object
using the estimated parameters and partition.
}
\examples{
\dontrun{
 ## the bull eyes data set
  data(bullsEye)
  model <- kmm( bullsEye[,1:2]
                        , 2
                        , models= "kmm_s"
                        )
  plot(model)
  }

}
\seealso{
\code{\link{plot}}
}
