% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmm.R
\docType{methods}
\name{plot,KmmComponent-method}
\alias{plot,KmmComponent-method}
\alias{plot-KmmComponent}
\title{Plotting of a class [\code{\linkS4class{KmmComponent}}]}
\usage{
\S4method{plot}{KmmComponent}(x, y, ...)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{KmmComponent}}]}

\item{y}{a vector with partitions}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting data from a [\code{\linkS4class{KmmComponent}}] object
using the estimated partition.
}
\examples{
\dontrun{
 ## the bull eyes data set
  data(bullsEye)
  model <- kmm( bullsEye, 2, models= "kmm_pk_s")
  plot(model)
  }

}
\seealso{
\code{\link{plot}}
}
