% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAlgo.R, R/ClusterAlgoPredict.R,
%   R/ClusterInit.R, R/ClusterStrategy.R, R/IClusterModel.R, R/global.R,
%   R/ClusterCategorical.R, R/ClusterDiagGaussian.R, R/ClusterGamma.R,
%   R/ClusterMixedData.R, R/ClusterPoisson.R, R/IClusterPredict.R,
%   R/ClusterPredict.R, R/LearnAlgo.R, R/kmm.R, R/kmmMixedData.R
\docType{methods}
\name{show,ClusterAlgo-method}
\alias{show,ClusterAlgo-method}
\alias{show}
\alias{show-algo,ClusterAlgo-method}
\alias{show,ClusterAlgoPredict-method}
\alias{show-algo,ClusterAlgoPredict-method}
\alias{show,ClusterInit-method}
\alias{show-init,ClusterInit,ClusterInit-method}
\alias{show,ClusterStrategy-method}
\alias{show-strategy,ClusterStrategy-method}
\alias{show,IClusterComponent-method}
\alias{show-IClusterComponent,IClusterComponent,IClusterComponent-method}
\alias{show,IClusterModel-method}
\alias{show,ClusterCategoricalComponent-method}
\alias{show-ClusterCategoricalComponent,ClusterCategoricalComponent,ClusterCategoricalComponent-method}
\alias{show,ClusterCategorical-method}
\alias{show-ClusterCategorical,ClusterCategorical,ClusterCategorical-method}
\alias{show,ClusterDiagGaussianComponent-method}
\alias{show-ClusterDiagGaussianComponent,ClusterDiagGaussianComponent,ClusterDiagGaussianComponent-method}
\alias{show,ClusterDiagGaussian-method}
\alias{show-ClusterDiagGaussian,ClusterDiagGaussian,ClusterDiagGaussian-method}
\alias{show,ClusterGammaComponent-method}
\alias{show-ClusterGammaComponent,ClusterGammaComponent,ClusterGammaComponent-method}
\alias{show,ClusterGamma-method}
\alias{show-ClusterGamma,ClusterGamma,ClusterGamma-method}
\alias{show,ClusterMixedDataModel-method}
\alias{show-ClusterMixedDataModel,ClusterMixedDataModel,ClusterMixedDataModel-method}
\alias{show,ClusterPoissonComponent-method}
\alias{show-ClusterPoissonComponent,ClusterPoissonComponent,ClusterPoissonComponent-method}
\alias{show,ClusterPoisson-method}
\alias{show-ClusterPoisson,ClusterPoisson,ClusterPoisson-method}
\alias{show,IClusterPredict-method}
\alias{show,ClusterPredict-method}
\alias{show,ClusterPredictMixedData-method}
\alias{show,LearnAlgo-method}
\alias{show-algo,LearnAlgo-method}
\alias{show,KmmComponent-method}
\alias{show-KmmComponent,KmmComponent,KmmComponent-method}
\alias{show,KmmModel-method}
\alias{show-KmmModel,KmmModel,KmmModel-method}
\alias{show,KmmMixedDataModel-method}
\alias{show-KmmMixedDataModel,KmmMixedDataModel,KmmMixedDataModel-method}
\title{Show description of a MixAll S4 class to standard output.}
\usage{
\S4method{show}{ClusterAlgo}(object)

\S4method{show}{ClusterAlgoPredict}(object)

\S4method{show}{ClusterInit}(object)

\S4method{show}{ClusterStrategy}(object)

\S4method{show}{IClusterComponent}(object)

\S4method{show}{IClusterModel}(object)

\S4method{show}{ClusterCategoricalComponent}(object)

\S4method{show}{ClusterCategorical}(object)

\S4method{show}{ClusterDiagGaussianComponent}(object)

\S4method{show}{ClusterDiagGaussian}(object)

\S4method{show}{ClusterGammaComponent}(object)

\S4method{show}{ClusterGamma}(object)

\S4method{show}{ClusterMixedDataModel}(object)

\S4method{show}{ClusterPoissonComponent}(object)

\S4method{show}{ClusterPoisson}(object)

\S4method{show}{IClusterPredict}(object)

\S4method{show}{ClusterPredict}(object)

\S4method{show}{ClusterPredictMixedData}(object)

\S4method{show}{LearnAlgo}(object)

\S4method{show}{KmmComponent}(object)

\S4method{show}{KmmModel}(object)

\S4method{show}{KmmMixedDataModel}(object)
}
\arguments{
\item{object}{a MixAll object: a \code{\linkS4class{ClusterStrategy}},
a \code{\linkS4class{ClusterInit}} or a \code{\linkS4class{ClusterAlgo}}.}
}
\value{
NULL. Prints to standard out.
}
\description{
Show description of a MixAll S4 class to standard output.
}
\examples{
  ## for strategy
  strategy <- clusterStrategy()
  show(strategy)
  ## for cluster init
  init <- clusterInit()
  show(init)
  ## for cluster algo
  algo <- clusterAlgo()
  show(algo)

}
\seealso{
\code{\link{show}}
}
