% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterGamma.R
\docType{class}
\name{ClusterGammaComponent}
\alias{ClusterGammaComponent}
\alias{ClusterGammaComponent-class}
\title{Definition of the [\code{\linkS4class{ClusterGammaComponent}}] class}
\description{
This class defines a gamma component of a mixture Model. It inherits
from [\code{\linkS4class{IClusterComponent}}].
}
\section{Slots}{

\describe{
\item{\code{shape}}{Matrix with the shapes of the jth variable in the kth cluster.}

\item{\code{scale}}{Matrix with the scales of the jth variable in the kth cluster.}
}}

\examples{
getSlots("ClusterGammaComponent")

}
\seealso{
[\code{\linkS4class{IClusterComponent}}] class
}
\author{
Serge Iovleff
}
