% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterPredict.R
\docType{class}
\name{ClusterPredict}
\alias{ClusterPredict}
\alias{ClusterPredict-class}
\title{Class [\code{\linkS4class{ClusterPredict}}] for predicting}
\description{
This class encapsulate the parameters for predicted data.
}
\section{Slots}{

\describe{
\item{\code{data}}{Matrix with the data set}

\item{\code{missing}}{Matrix with the indexes of the missing values}
}}

\examples{
  getSlots("ClusterPredict")

}
\seealso{
[\code{\linkS4class{IClusterPredict}}] class
}
\author{
Serge Iovleff
}
