% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IClusterModel.R, R/ClusterPoisson.R
\docType{class}
\name{IClusterComponent}
\alias{IClusterComponent}
\alias{IClusterComponent-class}
\alias{ClusterPoissonComponent}
\alias{ClusterPoissonComponent-class}
\title{Definition of the [\code{\linkS4class{IClusterComponent}}] class}
\description{
Interface base class defining a component of a mixture Model

This class defines a poisson component of a mixture Model. It inherits
from [\code{\linkS4class{IClusterComponent}}].
}
\section{Slots}{

\describe{
\item{\code{data}}{Matrix with the data set}

\item{\code{missing}}{Matrix with the indexes of the missing values}

\item{\code{modelName}}{model name associated with the data set}

\item{\code{lambda}}{Matrix with the mean of the jth variable in the kth cluster.}
}}

\examples{
getSlots("IClusterComponent")

getSlots("ClusterPoissonComponent")

}
\seealso{
[\code{\linkS4class{IClusterComponent}}] class
}
\author{
Serge Iovleff

Serge Iovleff
}
