% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MixAll.R
\docType{package}
\name{MixAll}
\alias{MixAll}
\alias{MixAll-package}
\title{MixAll allows to estimate parametric mixture models with heterogeneous data
set and missing data.}
\description{
This package contains methods allowing R users to use the clustering methods
of the STK++ library.
}
\details{
As described at the STK++ project's home page, \url{http://www.stkpp.org},
STK++ is a versatile, fast, reliable and elegant collection of C++ classes
for statistics, clustering, linear algebra, arrays (with an Eigen-like API),
regression, dimension reduction, etc. Some functionalities provided by the
library are available in the R environment as R functions in MixAll.

The available functionalities are:
\enumerate{
   \item the clusterDiagGaussian method allowing to discover group structures in
multivariate data sets using diagonal Gaussian multidimensional models (8 models).
   \item the clusterGamma method allowing to discover group structures in
multivariate data sets using gamma multidimensional models (24 models).
   \item the clusterPoisson method allowing to discover group structures in
multivariate data sets using poisson multidimensional models (8 models).
   \item the clusterHeterogeneous method allowing to discover group structures in
heterogeneous data sets.
 }

\tabular{ll}{
  Package: \tab rtkpp\cr
  Type: \tab Package\cr
  Version: \tab 0.8.1\cr
  Date: \tab 2014-07-05\cr
  License: \tab GPL for the MixAll and rtkpp side, LGPL for the stkpp side  + file LICENSE\cr
  LazyLoad: \tab yes\cr
}
}
\author{
Author: Serge Iovleff \email{contact@stkpp.org}
}
\keyword{STK++,}
\keyword{stkpp}

