% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterKernel.R
\docType{methods}
\name{plot,ClusterKernel-method}
\alias{plot,ClusterKernel-method}
\alias{plot-ClusterKernel}
\title{Plotting of a class [\code{\linkS4class{ClusterKernel}}]}
\usage{
\S4method{plot}{ClusterKernel}(x, y, ...)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{ClusterKernel}}]}

\item{y}{a list of variables to plot (subset). Variables names or indices.
If missing all the variables are represented.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting data from a [\code{\linkS4class{ClusterKernel}}] object
using the estimated parameters and partition.
}
\examples{
\dontrun{
 ## the bull eyes data set
  data(bullsEye)
  model <- clusterKernel( bullsEye[,1:2]
                        , 2
                        , models= "kernelGaussian_pk_s"
                        )
  plot(model)
  }

}
\seealso{
\code{\link{plot}}
}

