% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAlgo.R, R/ClusterInit.R, R/ClusterStrategy.R, R/IClusterModel.R, R/global.R, R/ClusterCategorical.R, R/ClusterDiagGaussian.R, R/ClusterGamma.R, R/ClusterKernel.R, R/ClusterMixedData.R, R/ClusterPoisson.R, R/LearnAlgo.R
\docType{methods}
\name{show,ClusterAlgo-method}
\alias{show}
\alias{show,ClusterAlgo-method}
\alias{show,ClusterCategorical-method}
\alias{show,ClusterCategoricalComponent-method}
\alias{show,ClusterDiagGaussian-method}
\alias{show,ClusterDiagGaussianComponent-method}
\alias{show,ClusterGamma-method}
\alias{show,ClusterGammaComponent-method}
\alias{show,ClusterInit-method}
\alias{show,ClusterKernel-method}
\alias{show,ClusterKernelComponent-method}
\alias{show,ClusterMixedData-method}
\alias{show,ClusterPoisson-method}
\alias{show,ClusterPoissonComponent-method}
\alias{show,ClusterStrategy-method}
\alias{show,IClusterComponent-method}
\alias{show,IClusterModel-method}
\alias{show,LearnAlgo-method}
\alias{show-ClusterCategorical,ClusterCategorical,ClusterCategorical-method}
\alias{show-ClusterCategoricalComponent,ClusterCategoricalComponent,ClusterCategoricalComponent-method}
\alias{show-ClusterDiagGaussian,ClusterDiagGaussian,ClusterDiagGaussian-method}
\alias{show-ClusterDiagGaussianComponent,ClusterDiagGaussianComponent,ClusterDiagGaussianComponent-method}
\alias{show-ClusterGamma,ClusterGamma,ClusterGamma-method}
\alias{show-ClusterGammaComponent,ClusterGammaComponent,ClusterGammaComponent-method}
\alias{show-ClusterKernel,ClusterKernel,ClusterKernel-method}
\alias{show-ClusterKernelComponent,ClusterKernelComponent,ClusterKernelComponent-method}
\alias{show-ClusterMixedData,ClusterMixedData,ClusterMixedData-method}
\alias{show-ClusterPoisson,ClusterPoisson,ClusterPoisson-method}
\alias{show-ClusterPoissonComponent,ClusterPoissonComponent,ClusterPoissonComponent-method}
\alias{show-IClusterComponent,IClusterComponent,IClusterComponent-method}
\alias{show-algo,ClusterAlgo-method}
\alias{show-algo,LearnAlgo-method}
\alias{show-init,ClusterInit,ClusterInit-method}
\alias{show-strategy,ClusterStrategy-method}
\title{Show description of a MixAll S4 class to standard output.}
\usage{
\S4method{show}{ClusterAlgo}(object)

\S4method{show}{ClusterInit}(object)

\S4method{show}{ClusterStrategy}(object)

\S4method{show}{IClusterModel}(object)

\S4method{show}{IClusterComponent}(object)

\S4method{show}{ClusterCategoricalComponent}(object)

\S4method{show}{ClusterCategorical}(object)

\S4method{show}{ClusterDiagGaussianComponent}(object)

\S4method{show}{ClusterDiagGaussian}(object)

\S4method{show}{ClusterGammaComponent}(object)

\S4method{show}{ClusterGamma}(object)

\S4method{show}{ClusterKernelComponent}(object)

\S4method{show}{ClusterKernel}(object)

\S4method{show}{ClusterMixedData}(object)

\S4method{show}{ClusterPoissonComponent}(object)

\S4method{show}{ClusterPoisson}(object)

\S4method{show}{LearnAlgo}(object)
}
\arguments{
\item{object}{a MixAll object: a \code{\linkS4class{ClusterStrategy}},
a \code{\linkS4class{ClusterInit}} or a \code{\linkS4class{ClusterAlgo}}.}
}
\value{
NULL. Prints to standard out.
}
\description{
Show description of a MixAll S4 class to standard output.
}
\examples{
  ## for strategy
  strategy <- clusterStrategy()
  show(strategy)
  ## for cluster init
  init <- clusterInit()
  show(init)
  ## for cluster algo
  algo <- clusterAlgo()
  show(algo)
}
\seealso{
\code{\link{show}}
}

