% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ClusterCategorical.R, R/ClusterDiagGaussian.R, R/ClusterGamma.R, R/ClusterHeterogeneous.R, R/ClusterKernel.R, R/ClusterPoisson.R, R/IClusterModel.R, R/global.R
\docType{methods}
\name{summary,ClusterCategorical-method}
\alias{summary}
\alias{summary,ClusterCategorical-method}
\alias{summary,ClusterDiagGaussian-method}
\alias{summary,ClusterGamma-method}
\alias{summary,ClusterHeterogeneous-method}
\alias{summary,ClusterKernel-method}
\alias{summary,ClusterPoisson-method}
\alias{summary,IClusterComponent-method}
\alias{summary,IClusterModelBase-method}
\alias{summary,MixmodResults-method}
\title{Produce summary of a MixAll S4 class.}
\usage{
\S4method{summary}{ClusterCategorical}(object, ...)

\S4method{summary}{ClusterDiagGaussian}(object, ...)

\S4method{summary}{ClusterGamma}(object, ...)

\S4method{summary}{ClusterHeterogeneous}(object, ...)

\S4method{summary}{ClusterKernel}(object, ...)

\S4method{summary}{ClusterPoisson}(object, ...)

\S4method{summary}{IClusterModelBase}(object, ...)

\S4method{summary}{IClusterComponent}(object, ...)
}
\arguments{
\item{object}{any cluster model deriving from a \code{\linkS4class{IClusterModelBase}} object.}

\item{...}{further arguments passed to or from other methods}
}
\value{
NULL. Summaries to standard out.
}
\description{
Produce summary of a MixAll S4 class.
}
\examples{
data(geyser)
  model <- clusterDiagGaussian(geyser,3)
  summary(model)
}
\seealso{
\code{\link{summary}}
}

