% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IClusterModel.R, R/global.R, R/ClusterCategorical.R, R/ClusterDiagGaussian.R, R/ClusterGamma.R, R/ClusterKernel.R, R/ClusterMixedData.R, R/ClusterPoisson.R
\docType{methods}
\name{summary,IClusterModel-method}
\alias{summary}
\alias{summary,ClusterCategorical-method}
\alias{summary,ClusterCategoricalComponent-method}
\alias{summary,ClusterDiagGaussian-method}
\alias{summary,ClusterGamma-method}
\alias{summary,ClusterKernel-method}
\alias{summary,ClusterMixedData-method}
\alias{summary,ClusterPoisson-method}
\alias{summary,IClusterComponent-method}
\alias{summary,IClusterModel-method}
\alias{summary-ClusterCategoricalComponent,ClusterCategoricalComponent,ClusterCategoricalComponent-method}
\title{Produce summary of a MixAll S4 class.}
\usage{
\S4method{summary}{IClusterModel}(object, ...)

\S4method{summary}{IClusterComponent}(object, ...)

\S4method{summary}{ClusterCategoricalComponent}(object)

\S4method{summary}{ClusterCategorical}(object, ...)

\S4method{summary}{ClusterDiagGaussian}(object, ...)

\S4method{summary}{ClusterGamma}(object, ...)

\S4method{summary}{ClusterKernel}(object, ...)

\S4method{summary}{ClusterMixedData}(object, ...)

\S4method{summary}{ClusterPoisson}(object, ...)
}
\arguments{
\item{object}{any cluster model deriving from a \code{\linkS4class{IClusterModel}}
or \code{\linkS4class{IClusterComponent}} object.}

\item{...}{further arguments passed to or from other methods}
}
\value{
NULL. Summaries to standard out.
}
\description{
Produce summary of a MixAll S4 class.
}

